# frontend/openapi.yml
openapi: 3.1.0
info:
  title: SpotPass Backend API
  description: Restaurant reservation and management system
  version: 1.0.0
paths:
...
  /api/staff/tables/:
    get:
      tags:
      - Staff - Tables
      summary: List Tables
      description: List all tables with optional filters (staff only)
      operationId: list_tables_api_staff_tables__get
      security:
      - HTTPBearer: []
      parameters:
      - name: zone_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Zone Id
      - name: is_available
        in: query
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Available
      - name: date_from
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Date From
      - name: date_to
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Date To
      - name: name
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TableRead'
                title: Response List Tables Api Staff Tables  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Staff - Tables
      summary: Create Table
      description: Create a new table (staff only)
      operationId: create_table_api_staff_tables__post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TableCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TableRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/staff/tables/{table_id}:
    get:
      tags:
      - Staff - Tables
      summary: Get Table
      description: Get a specific table (staff only)
      operationId: get_table_api_staff_tables__table_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: table_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Table Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TableRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - Staff - Tables
      summary: Update Table
      description: Update a table (staff only)
      operationId: update_table_api_staff_tables__table_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: table_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Table Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TableUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TableRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Staff - Tables
      summary: Delete Table
      description: Delete a table (staff only)
      operationId: delete_table_api_staff_tables__table_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: table_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Table Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/staff/tables/{table_id}/time-slots:
    get:
      tags:
      - Staff - Tables
      summary: Get Table Time Slots
      description: Get time slots for a table on a specific date (staff only)
      operationId: get_table_time_slots_api_staff_tables__table_id__time_slots_get
      security:
      - HTTPBearer: []
      parameters:
      - name: table_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Table Id
      - name: date
        in: query
        required: true
        schema:
          type: string
          format: date
          title: Date
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TimeSlotRead'
                title: Response Get Table Time Slots Api Staff Tables  Table Id  Time
                  Slots Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/staff/zones/:
    get:
      tags:
      - Staff - Zones
      summary: List Zones
      description: List all zones (staff only)
      operationId: list_zones_api_staff_zones__get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ZoneRead'
                type: array
                title: Response List Zones Api Staff Zones  Get
      security:
      - HTTPBearer: []
    post:
      tags:
      - Staff - Zones
      summary: Create Zone
      description: Create a new zone (staff only)
      operationId: create_zone_api_staff_zones__post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ZoneCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ZoneRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/staff/zones/{zone_id}:
    get:
      tags:
      - Staff - Zones
      summary: Get Zone
      description: Get a specific zone (staff only)
      operationId: get_zone_api_staff_zones__zone_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: zone_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Zone Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ZoneRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - Staff - Zones
      summary: Update Zone
      description: Update a zone (staff only)
      operationId: update_zone_api_staff_zones__zone_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: zone_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Zone Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ZoneUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ZoneRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Staff - Zones
      summary: Delete Zone
      description: Delete a zone (staff only)
      operationId: delete_zone_api_staff_zones__zone_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: zone_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Zone Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    EstablishmentRead:
      properties:
        name:
          type: string
          title: Name
        address:
          type: string
          title: Address
        id:
          type: string
          format: uuid
          title: Id
      type: object
      required:
      - name
      - address
      - id
      title: EstablishmentRead
      description: Schema for reading establishments
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    TableCreate:
      properties:
        name:
          type: string
          maxLength: 64
          minLength: 1
          title: Name
        description:
          anyOf:
          - type: string
            maxLength: 256
          - type: 'null'
          title: Description
        type:
          $ref: '#/components/schemas/TableType'
          default: table
        is_available:
          type: boolean
          title: Is Available
          default: true
        min_capacity:
          type: integer
          maximum: 100.0
          minimum: 1.0
          title: Min Capacity
        max_capacity:
          type: integer
          maximum: 100.0
          minimum: 1.0
          title: Max Capacity
        establishment_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Establishment Id
        zone_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Zone Id
      type: object
      required:
      - name
      - min_capacity
      - max_capacity
      title: TableCreate
      description: Schema for creating tables
      examples:
      - description: Window table with city view
        is_available: true
        max_capacity: 4
        min_capacity: 2
        name: Table 5
        type: table
        zone_id: 123e4567-e89b-12d3-a456-426614174001
...
    TableRead:
      properties:
        name:
          type: string
          maxLength: 64
          minLength: 1
          title: Name
        description:
          anyOf:
          - type: string
            maxLength: 256
          - type: 'null'
          title: Description
        type:
          $ref: '#/components/schemas/TableType'
          default: table
        is_available:
          type: boolean
          title: Is Available
          default: true
        min_capacity:
          type: integer
          maximum: 100.0
          minimum: 1.0
          title: Min Capacity
        max_capacity:
          type: integer
          maximum: 100.0
          minimum: 1.0
          title: Max Capacity
        id:
          type: string
          format: uuid
          title: Id
        created_at:
          type: string
          format: date-time
          title: Created At
        establishment:
          anyOf:
          - $ref: '#/components/schemas/EstablishmentRead'
          - type: 'null'
        zone:
          anyOf:
          - $ref: '#/components/schemas/ZoneRead'
          - type: 'null'
      type: object
      required:
      - name
      - min_capacity
      - max_capacity
      - id
      - created_at
      title: TableRead
      description: Schema for reading tables
      examples:
      - created_at: '2024-01-01T10:00:00'
        description: Window table with city view
        id: 123e4567-e89b-12d3-a456-426614174000
        is_available: true
        max_capacity: 4
        min_capacity: 2
        name: Table 5
        type: table
    TableType:
      type: string
      enum:
      - table
      - parasol
      - hut
      title: TableType
      description: Table type enumeration
    TableUpdate:
      properties:
        name:
          anyOf:
          - type: string
            maxLength: 64
            minLength: 1
          - type: 'null'
          title: Name
        description:
          anyOf:
          - type: string
            maxLength: 256
          - type: 'null'
          title: Description
        type:
          anyOf:
          - $ref: '#/components/schemas/TableType'
          - type: 'null'
        is_available:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Available
        min_capacity:
          anyOf:
          - type: integer
            maximum: 100.0
            minimum: 1.0
          - type: 'null'
          title: Min Capacity
        max_capacity:
          anyOf:
          - type: integer
            maximum: 100.0
            minimum: 1.0
          - type: 'null'
          title: Max Capacity
        zone_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Zone Id
      type: object
      title: TableUpdate
      description: Schema for updating tables
      examples:
      - description: Premium window table
        is_available: true
        max_capacity: 6
        min_capacity: 2
        name: Table 5A
    TimeSlotRead:
      properties:
        time:
          type: string
          title: Time
        status:
          type: string
          title: Status
        reservation:
          anyOf:
          - $ref: '#/components/schemas/TimeSlotReservationInfo'
          - type: 'null'
      type: object
      required:
      - time
      - status
      title: TimeSlotRead
      description: Time slot availability information
      examples:
      - reservation:
          client_name: John Doe
          guests: 4
          reference: REF1234567890
        status: occupied
        time: '19:30'
      - status: available
        time: '20:00'
    TimeSlotReservationInfo:
      properties:
        reference:
          type: string
          title: Reference
        guests:
          type: integer
          title: Guests
        client_name:
          type: string
          title: Client Name
      type: object
      required:
      - reference
      - guests
      - client_name
      title: TimeSlotReservationInfo
      description: Reservation information for a time slot
      examples:
      - client_name: John Doe
        guests: 4
        reference: REF1234567890
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    ZoneCreate:
      properties:
        name:
          type: string
          maxLength: 64
          minLength: 1
          title: Name
        establishment_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Establishment Id
      type: object
      required:
      - name
      title: ZoneCreate
      description: Schema for creating zones
      examples:
      - name: Terrace
    ZoneRead:
      properties:
        name:
          type: string
          maxLength: 64
          minLength: 1
          title: Name
        id:
          type: string
          format: uuid
          title: Id
      type: object
      required:
      - name
      - id
      title: ZoneRead
      description: Schema for reading zones
    ZoneUpdate:
      properties:
        name:
          anyOf:
          - type: string
            maxLength: 64
            minLength: 1
          - type: 'null'
          title: Name
      type: object
      title: ZoneUpdate
      description: Schema for updating zones
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer